requirements.txt
asgiref==3.6.0
Django==4.2
django-adminplus==0.5
django-resized==1.0.2
django-tinymce==3.6.1
mysqlclient==2.1.1
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
